--- src/test/resources/org/broadinstitute/hellbender/tools/exome/create-pon-control-full.pcov.gt.singular_values
@@ -1,100 +1,0 @@
-0.100032559801306
-0.00113943482724853
-0.00112122571559707
-0.00110811929923963
-0.00110639392006584
-0.00109348508239962
-0.00107939275249528
-0.00107437566345918
-0.00106497391363140
-0.00106224605454355
-0.00105768705597378
-0.00104852668044296
-0.00104367551194287
-0.00104223212799415
-0.00103730944903738
-0.00103405546982837
-0.00103108253020378
-0.00102478349675269
-0.00102253344842865
-0.00102150005044258
-0.00101843470508882
-0.00101393454923542
-0.00101115819790761
-0.00100963773056483
-0.00100763823900366
-0.00100483649436403
-0.00100152668701769
-0.000999404717822971
-0.000997269510156220
-0.000995599168783580
-0.000994091511764167
-0.000988452508494247
-0.000988265807566310
-0.000985303516150032
-0.000983659577290326
-0.000980656534106713
-0.000980242722071133
-0.000976705289405906
-0.000975658113691518
-0.000973131847864280
-0.000969867900344615
-0.000967553607940499
-0.000962880271315217
-0.000962008431028854
-0.000960618142951975
-0.000958030080949555
-0.000955419407161440
-0.000954306429443852
-0.000953134060443256
-0.000948039975786087
-0.000947066953223245
-0.000944910537392583
-0.000941724556054470
-0.000939401537548271
-0.000936486841970123
-0.000935075180125257
-0.000933105601977319
-0.000928866318907089
-0.000926410852421086
-0.000924145567921402
-0.000922277056588619
-0.000918833391769801
-0.000918127693942003
-0.000915647640366448
-0.000913926846757953
-0.000913261405597910
-0.000910137308557191
-0.000908170686978507
-0.000907167093454020
-0.000904201340490221
-0.000901654346033446
-0.000898595803675116
-0.000896902512521557
-0.000894711963412831
-0.000892144333557426
-0.000890104943341320
-0.000889349577960988
-0.000888100824620704
-0.000884332786915256
-0.000883164934997974
-0.000880536127825440
-0.000879182204807191
-0.000877292293348655
-0.000875231459653422
-0.000873076265153227
-0.000869012310354606
-0.000867178417872962
-0.000862618970943647
-0.000860799706170307
-0.000858132315834473
-0.000855512506485069
-0.000853081807962110
-0.000851271774134587
-0.000848052253721632
-0.000844153494566127
-0.000843483533469797
-0.000838016465365932
-0.000834530226913602
-0.000828279307643338
-0.000818837175718670
